Rename pages
This action will allow us to move some files that are not pages
inside the `pages` folder.
