+++ src/pages/cookies/index.page.jsx
