+++ src/pages/cursos/index.page.jsx
