+++ src/pages/terminos-legales/index.page.jsx
